Use memo and useCallback to stop unnecessary re-rendering
The pokemon card elements are constantly re-rendering whenever the gallery is re-rendering. This is because the handlePokemonSelect function that we pass in to set the selectedPokemonUrl state is changing with every render. I used the useCallback hook and wrapped it around the passed in function as well as memo() to wrap around the pokemon card components to stop them from being rerendered  upon every click